collapse hooks module
